% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bpf_loglike_sv}
\alias{bpf_loglike_sv}
\title{Implementing PMMH for the toy SV model.}
\usage{
bpf_loglike_sv(measurements, lNumber, starting_vals)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or measurements)}

\item{lNumber}{number of particles}

\item{starting_vals}{arma::vec giving the three starting values for phi_x,
sigma_x and beta_y}
}
\value{
double; value of the log-likelihood estimated via the BPF
}
\description{
The main cpp-function powering the R wrapper.
}
