% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pmmh_r.R
\name{bpf_pmmh}
\alias{bpf_pmmh}
\title{Boostrap Particle Filter for the toy SV model}
\usage{
bpf_pmmh(NN, yt, phi_x, sigma_x, beta_y)
}
\arguments{
\item{NN}{number of particles}

\item{yt}{measurements/data of the unobservered latent state process}

\item{phi_x}{phi parameter (see model equations for details)}

\item{sigma_x}{sigma parameter (see model equations for details)}

\item{beta_y}{beta parameter (see model equations for details)}
}
\value{
a list of one element (was previously returniong more) giving the 
  log-likelihood estimate
}
\description{
Computes, besides other things, an estimate of the log-likelihood.
}
