% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_simulation_helpers.R
\name{generate_data_simul_sv}
\alias{generate_data_simul_sv}
\title{Helper function simulating from the toy SV model}
\usage{
generate_data_simul_sv(TT, phi_x, sigma_x, beta_y, init_state_x0)
}
\arguments{
\item{TT}{number of time periods}

\item{phi_x}{phi parameter of state process}

\item{sigma_x}{sigma parameter of measurements y}

\item{beta_y}{beta parameter of measurements}

\item{init_state_x0}{state initialization}
}
\value{
a named list of two: simulated state trajectory (states_xt) and
 corresponding measurements (measurements_yt) under the model
}
\description{
Helper function simulating from the toy SV model
}
