% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_simulation_helpers.R
\name{generateDataSimulSV}
\alias{generateDataSimulSV}
\title{Helper function simulating from the toy SV model}
\usage{
generateDataSimulSV(TT, phiX, sigmaX, betaY, initStateX0)
}
\arguments{
\item{TT}{number of time periods}

\item{phiX}{phi parameter of state process}

\item{sigmaX}{sigma parameter of measurements y}

\item{betaY}{beta parameter of measurements}

\item{initStateX0}{state initialization}
}
\value{
a named list of two: simulated state trajectory (statesXt) and
 corresponding measurements (measurementsYt) under the model
}
\description{
Helper function simulating from the toy SV model
}
