% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_A_pg_r.R
\name{pg_sv}
\alias{pg_sv}
\title{Runs a particle Gibbs procedure on the toy SV model}
\usage{
pg_sv(
  y,
  Nparticles,
  starting_values,
  Niter,
  XcondInit,
  num_progress_outputs = 10
)
}
\arguments{
\item{y}{measurements i.e. observations of latent state process}

\item{Nparticles}{number of particles}

\item{starting_values}{vector of three giving the starting values for the 
parameters in the following order: phi_x, sigma_x, and beta_y}

\item{Niter}{number of PG iterations}

\item{XcondInit}{initial particle (trajectory) to condition on}
}
\value{

}
\description{
Runs a particle Gibbs procedure on the toy SV model
}
