% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_B_pgas_r.R
\name{pgas}
\alias{pgas}
\title{Title}
\usage{
pgas(M, y, Z, par_prior, N, par_inits)
}
\arguments{
\item{M}{}

\item{y}{}

\item{Z}{}

\item{par_prior}{}

\item{N}{}

\item{par_inits}{}
}
\value{

}
\description{
Title
}
