% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{plot_boxplot_loglike_estimates}
\alias{plot_boxplot_loglike_estimates}
\title{Generates a plot of boxplots of repeated log-likelihood BPF estimates}
\usage{
plot_boxplot_loglike_estimates(out_log_like, par_seq_phi, num_simul)
}
\arguments{
\item{out_log_like}{a matrix of dimension \code{length(par_seq_phi)} (rows)
times \code{num_simul}}

\item{par_seq_phi}{parameter sequence for phi parameter for which a 
log-likelihood estimate is computed}

\item{num_simul}{number of BPF runs per element of par_seq_phi}
}
\value{
invisible return; pure side effect of a boxplot generation
}
\description{
Loglikelihood should be obtained via a previous BPF run, that estimates
the SV model likelihood. These are passed, along with a sequence for the 
autoregressive parameter phi, the number of simulations (for each individual)
boxplot.
}
