% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{plot_pmcmc_output}
\alias{plot_pmcmc_output}
\title{Helper function for plotting PMMH and PGAS estimation results}
\usage{
plot_pmcmc_output(output_pmcmm, burnin, true_vals = NULL)
}
\arguments{
\item{burnin}{burnin period}

\item{true_vals}{defaults to NULL e.g. when a real dataset is used; for simulated data
can be set to true parameter values}

\item{output_pmcmc}{output object as provided by the pmmh estimatio functions}
}
\value{
invisible return; function called for plot side effects
}
\description{
Helper function for plotting PMMH and PGAS estimation results
}
