% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{plot_pmmh_output}
\alias{plot_pmmh_output}
\title{Helper function for plotting PMMH and PGAS estimation results}
\usage{
plot_pmmh_output(output_pmmh, burnin, true_vals = NULL, plot_type = "ilya")
}
\arguments{
\item{output_pmmh}{output object as provided by the pmmh estimatio functions}

\item{burnin}{burnin period}

\item{true_vals}{defaults to NULL e.g. when a real dataset is used; for simulated data
can be set to true parameter values}
}
\value{
invisible return; function called for plot side effects
}
\description{
Helper function for plotting PMMH and PGAS estimation results
}
