% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{plotBoxplotLogLikeEstimates}
\alias{plotBoxplotLogLikeEstimates}
\title{Generates a plot of boxplots of repeated log-likelihood BPF estimates}
\usage{
plotBoxplotLogLikeEstimates(outLogLike, parSeqPhi, numSimul)
}
\arguments{
\item{outLogLike}{a matrix of dimension \code{length(parSeqPhi)} (rows)
times \code{numSimul}}

\item{parSeqPhi}{parameter sequence for phi parameter for which a 
log-likelihood estimate is computed}

\item{numSimul}{number of BPF runs per element of parSeqPhi}
}
\value{
invisible return; pure side effect of a boxplot generation
}
\description{
Loglikelihood should be obtained via a previous BPF run, that estimates
the SV model likelihood. These are passed, along with a sequence for the 
autoregressive parameter phi, the number of simulations (for each individual)
boxplot.
}
