% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_plot_helpers.R
\name{plotPMCMCoutput}
\alias{plotPMCMCoutput}
\title{Helper function for plotting PMMH and PGAS estimation results}
\usage{
plotPMCMCoutput(outputPMCMC, burnin, trueVals = NULL)
}
\arguments{
\item{outputPMCMC}{output object as provided by the pmmh estimatio functions}

\item{burnin}{burnin period}

\item{trueVals}{defaults to NULL e.g. when a real dataset is used; for simulated data
can be set to true parameter values}
}
\value{
invisible return; function called for plot side effects
}
\description{
Helper function for plotting PMMH and PGAS estimation results
}
