% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pmmh_r.R
\name{pmmh_sv}
\alias{pmmh_sv}
\title{Runs a Particle Marginal Metropolis Hastings on the toy SV model}
\usage{
pmmh_sv(
  num_particles,
  rw_vcm_prop,
  y,
  starting_vals,
  num_iter,
  num_progress_outputs = 10
)
}
\arguments{
\item{num_particles}{number of particles}

\item{rw_vcm_prop}{variance-covariance matrix of the RW-MH proposal}

\item{y}{measurements}

\item{starting_vals}{vector of starting values for phi, sigma_x, and beta_y}

\item{num_iter}{number of PMMH iterations}
}
\value{
a list of parameter samples of for sigma-y and beta-y
}
\description{
Runs a Particle Marginal Metropolis Hastings on the toy SV model
}
