% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNonlin.R
\name{simNonlin}
\alias{simNonlin}
\title{simple nonlinear model}
\usage{
simNonlin(
  len = 50,
  var_init = 10,
  var_evol = 10,
  var_obs = 1,
  cosSeqOffset = -1
)
}
\arguments{
\item{len}{length}

\item{var_init}{initials}

\item{var_evol}{evol}

\item{var_obs}{obs}

\item{cosSeqOffset}{offset value}
}
\value{
data from the simple model
}
\description{
simple nonlinear model
}
