% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_AL_testing.R
\name{sv_model_al_tracking}
\alias{sv_model_al_tracking}
\title{Runs PMMH procedure for the toy SV model}
\usage{
sv_model_al_tracking(data, starting_vals, particles = 10, resample_freq = 10)
}
\arguments{
\item{data}{measurements or y}

\item{starting_vals}{starting values as a vector in the follwoing order
phi_x, sigma_x, and beta_y}

\item{particles}{number of particles}

\item{resample_freq}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
returns PMMH output
}
\description{
Runs PMMH procedure for the toy SV model
}
