% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sv_model_al_tracking_impl}
\alias{sv_model_al_tracking_impl}
\title{Implementing PMMH for the toy SV model.}
\usage{
sv_model_al_tracking_impl(
  measurements,
  starting_vals,
  lNumber,
  resample_freq = 0.5
)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or
measurements)}

\item{starting_vals}{arma::vec giving the three starting values for phi_x,
sigma_x and beta_y}

\item{lNumber}{number of particles}

\item{resample_freq}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples (sigma_x, 
  beta_y) and log-prior and log-likelihoood estimates
}
\description{
The main cpp-function powering the R wrapper.
}
