% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_model_pmmh.R
\name{sv_model_pmmh}
\alias{sv_model_pmmh}
\title{Runs PMMH procedure for the toy SV model}
\usage{
sv_model_pmmh(
  data,
  starting_vals,
  rw_mh_sd,
  particles = 2000,
  iterations = 5000,
  burnin = round(iterations/2, digits = 0),
  num_progress_outputs = 10
)
}
\arguments{
\item{data}{measurements or y}

\item{starting_vals}{starting values as a vector in the follwoing order
phi_x, sigma_x, and beta_y}

\item{rw_mh_sd}{a vector of two elements giving the standard deviations for the RW-MH step}

\item{particles}{number of particles}

\item{iterations}{number of PMCMC iterations}

\item{burnin}{burnin period (number of samples to exclude)}

\item{num_progress_outputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
returns PMMH output
}
\description{
Runs PMMH procedure for the toy SV model
}
