% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sv_model_pmmh_cpp}
\alias{sv_model_pmmh_cpp}
\title{Implementing PMMH for the toy SV model}
\usage{
sv_model_pmmh_cpp(data, lNumber, lMCMCits)
}
\arguments{
\item{data}{arma::vec providing the measurements (or y or data)}

\item{lNumber}{number of particles}

\item{lMCMCits}{number of PMMH iterations}
}
\value{
Rcpp::DataFrame containing the results: parameter samples and
  log-prior and log-likelihoood estimates
}
\description{
the main cpp-function powering the R wrapper
}
