% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svModelBpfLogLike}
\alias{svModelBpfLogLike}
\title{Implementing PMMH for the toy SV model.}
\usage{
svModelBpfLogLike(measurements, lNumber, initVals)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or measurements)}

\item{lNumber}{number of particles}

\item{initVals}{arma::vec giving the three starting values for phiX,
sigmaX and betaY}
}
\value{
double; value of the log-likelihood estimated via the BPF
}
\description{
The main cpp-function powering the R wrapper.
}
