% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_B_pg_cpp.R
\name{svModelPG}
\alias{svModelPG}
\title{Runs PG procedure for the toy SV model}
\usage{
svModelPG(
  data,
  startingVals,
  startingTrajectory,
  particles = 2000,
  iterations = 5000,
  numProgressOutputs = 10
)
}
\arguments{
\item{data}{measurements or y}

\item{startingVals}{starting values as a vector in the following order
phiX, sigmaX, and betaY}

\item{startingTrajectory}{initial reference trajectory condition for the
conditional SMC part of the PG algorith}

\item{particles}{number of particles}

\item{iterations}{number of PMCMC iterations}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples (sigmaX, betaY)
}
\description{
Runs PG procedure for the toy SV model
}
