% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmodelPGimpl}
\alias{svmodelPGimpl}
\title{Implementing PG for the toy SV model.}
\usage{
svmodelPGimpl(
  measurements,
  lNumber,
  lMCMCits,
  initVals,
  initReferenceTrajectory,
  num_progress_outputs = 10L
)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or measurements)}

\item{lNumber}{number of particles}

\item{lMCMCits}{number of PG iterations}

\item{initVals}{arma::vec giving the three starting values for phiX,
sigmaX and betaY}

\item{initReferenceTrajectory}{standard deviations for the RW-MH proposal step}

\item{num_progress_outputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples (sigmaX, 
  betaY) and log-prior and log-likelihoood estimates
}
\description{
The main cpp-function powering the R wrapper.
}
