% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_A_pg_r.R
\name{svModelPGr}
\alias{svModelPGr}
\title{Runs a particle Gibbs procedure on the toy SV model}
\usage{
svModelPGr(
  y,
  numParticles,
  startingValues,
  numIter,
  xRinit,
  asSampling = FALSE,
  numProgressOutputs = 10
)
}
\arguments{
\item{y}{measurements i.e. observations of latent state process}

\item{numParticles}{number of particles}

\item{startingValues}{vector of three giving the starting values for the 
parameters in the following order: phiX, sigmaX, and betaY}

\item{numIter}{number of PG iterations}

\item{xRinit}{initial particle to condition on (reference trajectory)}

\item{asSampling}{logical; if TRUE, then ancestor sampling is performed}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples (sigmaX, betaY)
}
\description{
Runs a particle Gibbs procedure on the toy SV model
}
