% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pmmh_smctc.R
\name{svModelPMMH}
\alias{svModelPMMH}
\title{Runs PMMH procedure for the toy SV model}
\usage{
svModelPMMH(
  data,
  initVals,
  rwMHsd,
  particles = 2000,
  iterations = 5000,
  burnin = round(iterations/2, digits = 0),
  numProgressOutputs = 10
)
}
\arguments{
\item{data}{measurements or y}

\item{initVals}{starting values as a vector in the follwoing order
phiX, sigmaX, and betaY}

\item{rwMHsd}{a vector of two elements giving the standard deviations for the RW-MH step}

\item{particles}{number of particles}

\item{iterations}{number of PMCMC iterations}

\item{burnin}{burnin period (number of samples to exclude)}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
returns PMMH output
}
\description{
Runs PMMH procedure for the toy SV model
}
