% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svModelPMMHimpl}
\alias{svModelPMMHimpl}
\title{Implementing PMMH for the toy SV model.}
\usage{
svModelPMMHimpl(
  measurements,
  lNumber,
  lMCMCits,
  initVals,
  rwMHvar,
  numProgressOutputs = 10L
)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or measurements)}

\item{lNumber}{number of particles}

\item{lMCMCits}{number of PMMH iterations}

\item{initVals}{arma::vec giving the three starting values for phiX,
sigmaX and betaY}

\item{rwMHvar}{standard deviations for the RW-MH proposal step}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples (sigmaX, 
  betaY) and log-prior and log-likelihoood estimates
}
\description{
The main cpp-function powering the R wrapper.
}
