% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pmmh_r.R
\name{svModelPMMHr}
\alias{svModelPMMHr}
\title{Runs a Particle Marginal Metropolis Hastings on the toy SV model}
\usage{
svModelPMMHr(
  numParticles,
  rwVCMprop,
  y,
  initVals,
  numIter,
  numProgressOutputs = 10
)
}
\arguments{
\item{numParticles}{number of particles}

\item{rwVCMprop}{variance-covariance matrix of the RW-MH proposal}

\item{y}{measurements}

\item{initVals}{vector of starting values for phi, sigmaX, and betaY}

\item{numIter}{number of PMMH iterations}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
a list of parameter samples of for sigma-y and beta-y
}
\description{
Runs a Particle Marginal Metropolis Hastings on the toy SV model
}
