% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVmodelRcppSMC.R
\docType{package}
\name{SVmodelRcppSMC}
\alias{SVmodelRcppSMC}
\title{SVmodelRcppSMC: A package for example implementations of PMCMC inference}
\description{
This SVmodelRcppSMC package provides Particle Gibbs and Particle Gibbs with
ancestor sampling estimation of the standard SV model using the more general
RcppSMC package for the conditional SMC part of the algorithm.
}
