% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_A_pg_r.R
\name{pg_sv}
\alias{pg_sv}
\title{Runs a particle Gibbs procedure on the toy SV model}
\usage{
pg_sv(
  y,
  num_particles,
  starting_values,
  num_iter,
  x_r_init,
  as_sampling,
  num_progress_outputs = 10
)
}
\arguments{
\item{y}{measurements i.e. observations of latent state process}

\item{num_particles}{number of particles}

\item{starting_values}{vector of three giving the starting values for the 
parameters in the following order: phi_x, sigma_x, and beta_y}

\item{num_iter}{number of PG iterations}

\item{x_r_init}{initial particle to condition on (reference trajectory)}
}
\value{

}
\description{
Runs a particle Gibbs procedure on the toy SV model
}
