% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_AL_testing.R
\name{sv_model_al_tracking}
\alias{sv_model_al_tracking}
\title{Runs PMMH procedure for the toy SV model}
\usage{
sv_model_al_tracking(data, starting_vals, particles = 10, resampleFreq = 10)
}
\arguments{
\item{data}{measurements or y}

\item{starting_vals}{starting values as a vector in the follwoing order
phi_x, sigma_x, and beta_y}

\item{particles}{number of particles}

\item{resampleFreq}{frequency at which resampling is performed; if negative,
then resampling is never performed; if between [0,1), then resampling is
performed when  the ESS falls below that proportion of the number of
particles and when it is greater than or equal to 1, resampling is carried
out when the ESS falls below that value (note: if this parameter is larger
than the total number of particles, then resampling will always be
performed!)}
}
\value{
returns PMMH output
}
\description{
Runs PMMH procedure for the toy SV model
}
