% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sv_model_al_tracking_impl}
\alias{sv_model_al_tracking_impl}
\title{Implementing PMMH for the toy SV model.}
\usage{
sv_model_al_tracking_impl(
  measurements,
  starting_vals,
  lNumber,
  resampleFreq = 0.5
)
}
\arguments{
\item{measurements}{arma::vec providing the measurements (or y or
measurements)}

\item{starting_vals}{arma::vec giving the three starting values for phi_x,
sigma_x and beta_y}

\item{lNumber}{number of particles}

\item{resampleFreq}{frequency at which resampling is performed; if
negative, then resampling is never performed; if between [0,1), then
resampling is performed when  the ESS falls below that proportion of the
number of particles and when it is greater than or equal to 1, resampling
is carried out when the ESS falls below that value (note: if this
parameter is larger than the total number of particles, then resampling
will always be performed!)}
}
\value{
Rcpp::List containing the results: parameter samples (sigma_x, 
  beta_y) and log-prior and log-likelihoood estimates
}
\description{
The main cpp-function powering the R wrapper.
}
