% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_model_pmmh.R
\name{sv_model_pmmh}
\alias{sv_model_pmmh}
\title{Runs PMMH procedure on the simulated data example}
\usage{
sv_model_pmmh(
  data,
  particles = 5000,
  iterations = 10000,
  burnin = 0,
  plot = FALSE
)
}
\arguments{
\item{data}{measurements or y}

\item{particles}{number of particles}

\item{iterations}{number of PMCMC iterations}

\item{burnin}{burnin period (number of samples to exclude)}

\item{plot}{logical; if TRUE, a plot of the reslults will be returned}
}
\value{
invisibly returns PMMH output and, if plot = TRUE, a plot of the
 output
}
\description{
Runs PMMH procedure on the simulated data example
}
