% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_B_pg_cpp.R
\name{svmodelPG}
\alias{svmodelPG}
\title{Runs PG procedure for the toy SV model}
\usage{
svmodelPG(
  data,
  startingVals,
  startingTrajectory,
  particles = 2000,
  iterations = 5000,
  burnin = round(iterations/2, digits = 0),
  numProgressOutputs = 10
)
}
\arguments{
\item{data}{measurements or y}

\item{startingVals}{starting values as a vector in the follwoing order
phi_x, sigma_x, and beta_y}

\item{particles}{number of particles}

\item{iterations}{number of PMCMC iterations}

\item{burnin}{burnin period (number of samples to exclude)}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
returns PG output
}
\description{
Runs PG procedure for the toy SV model
}
