% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svModelPGimpl}
\alias{svModelPGimpl}
\title{Implementing PG for the toy SV model.}
\usage{
svModelPGimpl(
  measurements,
  lNumber,
  lMCMCits,
  initVals,
  initReferenceTrajectory,
  numProgressOutputs = 10L
)
}
\arguments{
\item{measurements}{arma::vec providing the measurements}

\item{lNumber}{number of particles}

\item{lMCMCits}{number of PG iterations}

\item{initVals}{arma::vec giving the three starting values for phiX,
sigmaX and betaY}

\item{initReferenceTrajectory}{initial reference trajectory values to 
condition on}

\item{numProgressOutputs}{int giving the number of progress outputs i.e.
if set to 10, then progress output occurs for every additional 10% of 
completion}
}
\value{
Rcpp::List containing the results: parameter samples sigmaX, 
  betaY
}
\description{
The main cpp-function powering the R wrapper for the PG sampler.
}
